a placeholder map visualization
A placeholder map has been included using the Leaflet.js library. The app and the actions should still be renamed and the Dockerfile for the jsonAPI needs to be updated so that the correct files are added to the $CATALINA_HOME folder for the Apache Tomcat server we are running. The intent is to use simple shapes for the visualization of boats in the Boston Harbor. The postgres JDBC plugin will be used to pull data from the already running database. That code has yet to be written.